#refactoring: (refactor) changed webgpu structure
